Minor code review

+++ LoadTest/Game.cpp
@@ -320,7 +320,7 @@ void Game::Render()
         if (FAILED(hr))
         {
             char buff[128] = {};
-            sprintf_s(buff, "ERROR: SaveWICTextureToFile (PNG) failed %08X\n", hr);
+            sprintf_s(buff, "ERROR: SaveWICTextureToFile (PNG) failed %08X\n", static_cast<unsigned int>(hr));
             OutputDebugStringA(buff);
             success = false;
         }
@@ -341,7 +341,7 @@ void Game::Render()
         if (FAILED(hr))
         {
             char buff[128] = {};
-            sprintf_s(buff, "ERROR: SaveWICTextureToFile (JPG) failed %08X\n", hr);
+            sprintf_s(buff, "ERROR: SaveWICTextureToFile (JPG) failed %08X\n", static_cast<unsigned int>(hr));
             OutputDebugStringA(buff);
             success = false;
         }
@@ -363,7 +363,7 @@ void Game::Render()
         if (FAILED(hr))
         {
             char buff[128] = {};
-            sprintf_s(buff, "ERROR: SaveWICTextureToFile (BMP) failed %08X\n", hr);
+            sprintf_s(buff, "ERROR: SaveWICTextureToFile (BMP) failed %08X\n", static_cast<unsigned int>(hr));
             OutputDebugStringA(buff);
             success = false;
         }
@@ -400,7 +400,7 @@ void Game::Render()
         if (FAILED(hr))
         {
             char buff[128] = {};
-            sprintf_s(buff, "ERROR: SaveWICTextureToFile (TIFF) failed %08X\n", hr);
+            sprintf_s(buff, "ERROR: SaveWICTextureToFile (TIFF) failed %08X\n", static_cast<unsigned int>(hr));
             OutputDebugStringA(buff);
             success = false;
         }
@@ -420,7 +420,7 @@ void Game::Render()
         if (FAILED(hr))
         {
             char buff[128] = {};
-            sprintf_s(buff, "ERROR: SaveDDSTextureToFile failed %08X\n", hr);
+            sprintf_s(buff, "ERROR: SaveDDSTextureToFile failed %08X\n", static_cast<unsigned int>(hr));
             OutputDebugStringA(buff);
             success = false;
         }
